requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
black==25.1.0
    # via -r requirements.in
catboost==1.2.8
    # via -r requirements.in
cfgv==3.4.0
    # via pre-commit
click==8.2.1
    # via black
contourpy==1.3.2
    # via matplotlib
coverage[toml]==7.8.2
    # via pytest-cov
cycler==0.12.1
    # via matplotlib
distlib==0.3.9
    # via virtualenv
filelock==3.18.0
    # via
    #   torch
    #   triton
    #   virtualenv
fonttools==4.58.1
    # via matplotlib
fsspec==2025.5.1
    # via torch
graphviz==0.20.3
    # via catboost
hdbscan==0.8.40
    # via -r requirements.in
identify==2.6.12
    # via pre-commit
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via torch
joblib==1.5.1
    # via
    #   -r requirements.in
    #   hdbscan
    #   pynndescent
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lightgbm==4.6.0
    # via -r requirements.in
llvmlite==0.44.0
    # via
    #   numba
    #   pynndescent
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via
    #   -r requirements.in
    #   catboost
    #   seaborn
mpmath==1.3.0
    # via sympy
mypy==1.16.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
narwhals==1.41.0
    # via plotly
networkx==3.3
    # via
    #   -r requirements.in
    #   torch
nodeenv==1.9.1
    # via pre-commit
numba==0.61.2
    # via
    #   pynndescent
    #   umap-learn
numpy==2.2.6
    # via
    #   -r requirements.in
    #   catboost
    #   contourpy
    #   hdbscan
    #   lightgbm
    #   matplotlib
    #   numba
    #   pandas
    #   pandas-stubs
    #   scikit-learn
    #   scipy
    #   seaborn
    #   umap-learn
    #   xgboost
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.19.3
    # via
    #   torch
    #   xgboost
nvidia-nvjitlink-cu12==12.9.41
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
packaging==25.0
    # via
    #   black
    #   matplotlib
    #   plotly
    #   pytest
pandas==2.2.3
    # via
    #   -r requirements.in
    #   catboost
    #   seaborn
pandas-stubs==2.2.3.250527
    # via -r requirements.in
pathspec==0.12.1
    # via
    #   black
    #   mypy
pillow==11.2.1
    # via matplotlib
platformdirs==4.3.8
    # via
    #   black
    #   virtualenv
plotly==6.1.2
    # via catboost
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via -r requirements.in
pydantic==2.11.5
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pynndescent==0.5.13
    # via umap-learn
pyparsing==3.2.3
    # via matplotlib
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==6.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pre-commit
ruff==0.11.12
    # via -r requirements.in
scikit-learn==1.6.1
    # via
    #   -r requirements.in
    #   hdbscan
    #   pynndescent
    #   umap-learn
scipy==1.15.3
    # via
    #   -r requirements.in
    #   catboost
    #   hdbscan
    #   lightgbm
    #   pynndescent
    #   scikit-learn
    #   umap-learn
    #   xgboost
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via
    #   catboost
    #   python-dateutil
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.2.0
    # via -r requirements.in
tqdm==4.67.1
    # via umap-learn
triton==2.2.0
    # via torch
types-pytz==2025.2.0.20250516
    # via
    #   -r requirements.in
    #   pandas-stubs
types-pyyaml==6.0.12.20250516
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   -r requirements.in
    #   mypy
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
umap-learn==0.5.7
    # via -r requirements.in
virtualenv==20.31.2
    # via pre-commit
xgboost==3.0.2
    # via -r requirements.in
